Fix miss GT line in region_layer
Change gt loop *break* to *continue*.
As we tested, some of the GT box will be change to 0 after the data augmentation. Thus, the following GT boxes will be ignored, if the x of this line is 0. The mAP validation of PASCAL VOC of this change will increase.

The following example of a miss GT (the third line is a miss GT line):
0.02 0.48 0.04 0.07 6
0.70 0.47 0.26 0.10 5
0.00 0.00 0.00 0.00 0
0.35 0.58 0.12 0.07 6
0.21 0.57 0.34 0.17 6
0.17 0.74 0.34 0.49 6
...